Adapt Sharp A903 for IVT CRMC-A768JBEZ
Most signals work out of the box, but there are some breaking
differences:
IVT has got both horizontal and vertical swing.
Vertical swing has got more allowed positions for the different modes.

Furthermore temperature is two degrees off, which is not compensated for. It must be handled in the application.